Reorganize zh_CN internet test
